improved the layout of why franchizerz
