--- src/components/WhyFranchizerz/WhyFranchizerz.module.css
@@ -1,110 +1,0 @@
-/* @import "https://unpkg.com/@picocss/pico@1.5.5/css/pico.min.css"; */
-
-@keyframes fadeIn {
-	0% {
-		opacity: 0;
-		transform: translateY(20px);
-	}
-	100% {
-		opacity: 1;
-		transform: translateY(0);
-	}
-}
-
-@keyframes pulse {
-	0%,
-	100% {
-		transform: scale(1);
-	}
-	50% {
-		transform: scale(1.05);
-	}
-}
-
-@keyframes backgroundAnimation {
-	0% {
-		background-position: 0% 50%;
-	}
-	50% {
-		background-position: 100% 50%;
-	}
-	100% {
-		background-position: 0% 50%;
-	}
-}
-
-.container1 {
-	/* background: linear-gradient(-45deg, #1d1f20, #3d4145, #1d1f20); */
-	background: #fff;
-	background-size: 400% 400%;
-	padding: 60px;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	/* color: #fff; */
-	color: rgb(241, 206, 162);
-	text-align: center;
-	overflow: hidden;
-	animation: fadeIn 1s ease-out, backgroundAnimation 15s ease infinite;
-}
-
-.heading1 {
-	font-size: 3rem;
-	margin-bottom: 30px;
-	animation: pulse 2s infinite;
-}
-
-.featuresGrid {
-	display: flex;
-	flex-wrap: wrap;
-	justify-content: space-around;
-	align-items: center;
-	gap: 30px;
-	width: 100%;
-	max-width: 1200px;
-}
-
-.featureCard {
-	background: rgba(59, 130, 246, 0.9);
-	backdrop-filter: blur(10px);
-	padding: 40px;
-	border-radius: 15px;
-	box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-	transition: transform 0.3s, box-shadow 0.3s;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	gap: 15px;
-	animation: fadeIn 0.5s ease-out forwards;
-	color: #e0e0e0;
-	width: 45%;
-}
-
-.featureCard:hover {
-	transform: translateY(-10px);
-	box-shadow: 0 8px 16px rgba(0, 0, 0, 0.3);
-	animation: pulse 0.5s;
-}
-
-.featureCard p {
-	font-size: 1.5rem;
-	margin: 0;
-	text-align: center;
-}
-
-.icon1 {
-	font-size: 2.5rem;
-	color: #23d0e9;
-	animation: rotate 3s infinite linear;
-}
-
-@media (max-width: 768px) {
-	.featuresGrid {
-		flex-direction: column;
-	}
-
-	.featureCard {
-		width: 80%;
-	}
-}
